--- simlify/structure/__init__.py
@@ -1,3 +1,0 @@
-from simlify.structure.extract import extract_atoms
-
-__all__ = ["extract_atoms"]
