+++ simlify/tasks/slice/extract.py
